% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiducialEstimator.R
\name{Fiducial}
\alias{Fiducial}
\title{Fiducial Estimaator}
\usage{
Fiducial(L, R, timepoints = NULL, timeupper = NULL, alpha = 0.95)
}
\arguments{
\item{L, R}{The interval censored data. L < R. It shoule be two numeric vectors.}

\item{timepoints}{The interesting timepoints you want to estimate. Default is all timepoints
from L,R}

\item{timeupper}{The upper bound of the test timepoints. Default is the max finite time of R.}

\item{alpha}{Significant level}
}
\value{
The estimated cumulative density
}
\description{
Estimate the interval censored data [L,R] to estimate the cdf function P[Y <= t]
}
